Replace `TrueTypeTableBuilder` and `CompilerOutput` with a single class
Given that both of these classes are so similar, let's replace them with a single `DataBuilder` class instead to reduce unnecessary code-duplication.
